Tail Statistic

Generate subintensity matrix, vector of initial probabilites and defect for combination of frequency counts higher than "k" (tail statistic)

tailstat(mph_obj, k, theta)

Arguments

mph_obj
Multivariate Phase Type object generated either from mult_phase_type or kingsman function

k
Minimum Frequency count (positive integer)

theta
mutation parameter (positive)

Value

A `disc_phase_type` object containing subintensity matrix (P), vector of initial probabilities (alpha) and defect (probability of not entering any transient state prior to absorption)

Examples

tailstat(kingsman(15), k = 10, theta = 2)

#> $subint_mat
#>         [,1]       [,2]       [,3]       [,4]       [,5]       [,6]       [,7]
#>  [1,] 0.0625 0.02840909 0.05681818 0.02435065 0.04870130 0.02435065 0.02435065
#>  [2,] 0.0000 0.09090909 0.00000000 0.05194805 0.07792208 0.00000000 0.00000000
#>  [3,] 0.0000 0.00000000 0.09090909 0.01298701 0.03896104 0.03896104 0.03896104
#>  [4,] 0.0000 0.00000000 0.00000000 0.14285714 0.00000000 0.00000000 0.00000000
#>  [5,] 0.0000 0.00000000 0.00000000 0.00000000 0.14285714 0.00000000 0.00000000
#>  [6,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.14285714 0.00000000
#>  [7,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.14285714
#>  [8,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [12,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [13,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [14,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [15,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.0000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             [,8]       [,9]      [,10]      [,11]      [,12]      [,13]
#>  [1,] 0.03652597 0.05478896 0.03652597 0.01826299 0.01826299 0.01826299
#>  [2,] 0.05844156 0.07792208 0.03896104 0.01948052 0.00000000 0.00000000
#>  [3,] 0.02922078 0.04870130 0.03896104 0.01948052 0.02922078 0.02922078
#>  [4,] 0.10714286 0.10714286 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.03571429 0.07142857 0.07142857 0.03571429 0.00000000 0.00000000
#>  [6,] 0.03571429 0.00000000 0.07142857 0.00000000 0.07142857 0.03571429
#>  [7,] 0.00000000 0.07142857 0.00000000 0.03571429 0.03571429 0.07142857
#>  [8,] 0.25000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.00000000 0.25000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.00000000 0.00000000 0.25000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.00000000 0.00000000 0.00000000 0.25000000 0.00000000 0.00000000
#> [12,] 0.00000000 0.00000000 0.00000000 0.00000000 0.25000000 0.00000000
#> [13,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.25000000
#> [14,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.09131494 0.07305195 0.05478896 0.03652597 0.01826299
#>  [2,] 0.11688312 0.08766234 0.05844156 0.02922078 0.00000000
#>  [3,] 0.08766234 0.07305195 0.05844156 0.04383117 0.02922078
#>  [4,] 0.16071429 0.10714286 0.05357143 0.00000000 0.00000000
#>  [5,] 0.10714286 0.08928571 0.07142857 0.05357143 0.00000000
#>  [6,] 0.10714286 0.03571429 0.05357143 0.07142857 0.05357143
#>  [7,] 0.05357143 0.10714286 0.07142857 0.03571429 0.05357143
#>  [8,] 0.25000000 0.12500000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.12500000 0.12500000 0.12500000 0.00000000 0.00000000
#> [10,] 0.12500000 0.00000000 0.12500000 0.12500000 0.00000000
#> [11,] 0.00000000 0.25000000 0.00000000 0.12500000 0.00000000
#> [12,] 0.12500000 0.00000000 0.00000000 0.12500000 0.12500000
#> [13,] 0.00000000 0.12500000 0.12500000 0.00000000 0.12500000
#> [14,] 0.50000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [15,] 0.00000000 0.50000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.00000000 0.00000000 0.50000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.50000000 0.00000000
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.50000000
#> 
#> $init_probs
#>             [,1]        [,2]       [,3]        [,4]       [,5]       [,6]
#> [1,] 0.002997003 0.003996004 0.01798202 0.006993007 0.02397602 0.02697303
#>            [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#> [1,] 0.02697303 0.01648352 0.03846154 0.04395604 0.02197802 0.04945055
#>           [,13]      [,14]      [,15]      [,16]      [,17]     [,18]
#> [1,] 0.04945055 0.05494505 0.06593407 0.07692308 0.08791209 0.0989011
#> 
#> $defect
#> [1] 0.2857143
#> 
#> attr(,"class")
#> [1] "disc_phase_type"